Refactor dataset service imports to use new knowledge service structure and remove deprecated use-datasets service. Update related components for consistency.
